\name{genRal-package}
\alias{genRal-package}
\alias{genRal}
\docType{package}
\title{
  Utilities for data analysis
}
\description{
  This package contains tools and helper functions to perform (usually small)
  manipulations of data structures such as data frames, matrices and lists,
  assist in gene expression data analysis and provide specific functionality
  to analysis scripts.
}
\details{
\tabular{ll}{
  Package: \tab genRal\cr
  Type: \tab Package\cr
  Version: \tab 0.1-0\cr
  License: \tab MIT License\cr
  LazyLoad: \tab yes\cr
}
}
\author{
  Vidal Fey <vidal.fey@utu.fi>

  Maintainer: Dhanaprakash Jambulingam <dhajam@utu.fi>
}
\keyword{ package }
